+++ src/functions/HammingDistance.h
@@ -49,7 +49,7 @@ public:
 
   // Return the value of c(X), where X is a subset.
   //
-  float cost (ElementSubset *);
+  double cost (ElementSubset *);
 
 };
